Refactor monetary values to use integer cents
- Introduced a new Money utility class for handling monetary arithmetic in integer cents to avoid floating-point precision issues.
- Updated all relevant tests to reflect changes in monetary value representation, converting decimal values to integer cents.
- Modified database migrations to convert existing monetary columns from decimal to integer, ensuring consistency across the application.
- Adjusted test cases across various feature and unit tests to accommodate the new monetary representation.
- Ensured that all monetary values are now stored and manipulated as integers, improving accuracy in billing and invoicing processes.

+++ app/Actions/Platform/EnableCorporateModuleAction.php
@@ -16,7 +16,10 @@ class EnableCorporateModuleAction
 {
     public function __construct(private readonly SubscriptionCalculator $calculator) {}
 
-    public function execute(Corporation $corporation, string $moduleCode, ?float $customMonthlyPrice = null): CorporationModule
+    /**
+     * @param  int|null  $customMonthlyPrice  Centavos.
+     */
+    public function execute(Corporation $corporation, string $moduleCode, ?int $customMonthlyPrice = null): CorporationModule
     {
         $code = ModuleCode::tryFrom($moduleCode);
